+++ public/data/CommentData.json
@@ -1,0 +1,14 @@
+[
+  {
+    "id": 1,
+    "text": "리액트 너무 어려워"
+  },
+  {
+    "id": 2,
+    "text": "Mock data 적용하기"
+  },
+  {
+    "id": 3,
+    "text": "React!"
+  }
+]
