+++ src/pages/jooyoungPark/Main/Main.js
@@ -1,118 +1,35 @@
-import React, { useState, useRef } from 'react';
-import Nav from '../../../components/Nav/Nav.js';
+import React, { useEffect, useState } from 'react';
+import { Link } from 'react-router-dom';
 import './Main.scss';
 import '../../../styles/variables.scss';
-import Commentbox from '../components/Commentbox.js';
-import { Link } from 'react-router-dom';
+import Nav from '../../../components/Nav/Nav.js';
 import Rightbox from '../components/Rightbox.js';
+import Feedbox from '../components/Feedbox';
 
 function Main() {
-  const [value, setValue] = useState('');
-  let [comment, setComment] = useState([]);
-
-  const onChange = e => {
-    setValue(e.target.value);
-  };
-
-  // useRef()로 관리하는 변수는 값이 바뀌어도 리렌더링 되지 않는다.
-  const nextId = useRef(0);
-  const onSubmit = e => {
-    setComment([
-      ...comment,
-      {
-        id: nextId.current,
-        text: value,
-      },
-    ]);
-
-    // input 초기화
-    setValue('');
-
-    // id값 1씩 증가
-    nextId.current += 1;
-
-    // 새로고침 X
-    e.preventDefault();
-  };
-
-  const onRemove = id => {
-    setComment(comment.filter(com => com.id !== id));
-  };
-
+  const [feed, setFeed] = useState([]);
+  useEffect(() => {
+    fetch('data/Feed.json')
+      .then(red => red.json())
+      .then(data => {
+        setFeed(data);
+      });
+  }, []);
   return (
     <div className="main">
       <Nav />
 
       <main>
         <div className="feed">
-          <article className="post">
-            <div className="user_post_wrap">
-              <div className="user_info">
-                <div className="user_img">
-                  <Link to="/main-jy">
-                    <img
-                      src="images/jooyoungPark/Main/mini_profile.jpg"
-                      alt="/main-jy"
-                    />
-                  </Link>
-                </div>
-                <div className="user_name">
-                  <Link to="/main-jy">cuteCat</Link>
-                  <Link to="/main-jy">귀여움이 세상을 지배한다.</Link>
-                </div>
-                <button className="more_btn">더보기</button>
-              </div>
-
-              <div className="user_post_img">
-                <img src="images/jooyoungPark/Main/cat.jpg" alt="" />
-              </div>
-
-              <div className="user_post_desc">
-                <div className="user_check_wrap">
-                  <div>
-                    <button>
-                      <img src="images/jooyoungPark/Main/heart.png" alt="" />
-                    </button>
-                    <button>
-                      <img src="images/jooyoungPark/Main/chat.png" alt="" />
-                    </button>
-                    <button>
-                      <img src="images/jooyoungPark/Main/send.png" alt="" />
-                    </button>
-                  </div>
-                  <button>북마크</button>
-                </div>
-                <div className="like_count">
-                  <Link to="/main-jy">
-                    <img
-                      src="images/jooyoungPark/Main/mini_profile.jpg"
-                      alt="/main-jy"
-                    />
-                  </Link>
-                  <span className="like_desc">
-                    lemon님 외 4명이 좋아합니다.
-                  </span>
-                </div>
-                <div className="user_post_content">
-                  <strong className="user_id">cuteCat</strong>
-                  <span>고양이 귀여워</span>
-                  <Commentbox comment={comment} onRemove={onRemove} />
-                  <button className="more_comment">댓글 147개 모두 보기</button>
-                </div>
-              </div>
-
-              <form className="comment_wrap" onSubmit={onSubmit}>
-                <input
-                  type="text"
-                  placeholder="댓글 달기..."
-                  required
-                  value={value}
-                  onChange={onChange}
-                />
-                <button className="comment_enter">게시</button>
-              </form>
-            </div>
-          </article>
+          {feed.map(f => (
+            <Feedbox
+              key={f.id}
+              userId={f.userId}
+              desc={f.desc}
+              profileImg={f.profileImg}
+              feedImg={f.feedImg}
+            />
+          ))}
         </div>
 
         <div className="main_right">
@@ -137,7 +54,7 @@ function Main() {
               <Link to="/main-jy">모두 보기</Link>
             </div>
             <ul>
-              <li className="user_progile_mini">
+              <li className="user_profile_mini">
                 <Link to="/main-jy">
                   <img
                     src="images/jooyoungPark/Main/mini_profile.jpg"
@@ -149,7 +66,7 @@ function Main() {
                   <p>2시간 전</p>
                 </div>
               </li>
-              <li className="user_progile_mini">
+              <li className="user_profile_mini">
                 <Link to="/main-jy">
                   <img
                     src="images/jooyoungPark/Main/mini_profile.jpg"
@@ -161,7 +78,7 @@ function Main() {
                   <p>2시간 전</p>
                 </div>
               </li>
-              <li className="user_progile_mini">
+              <li className="user_profile_mini">
                 <Link to="/main-jy">
                   <img
                     src="images/jooyoungPark/Main/mini_profile.jpg"
@@ -173,7 +90,7 @@ function Main() {
                   <p>2시간 전</p>
                 </div>
               </li>
-              <li className="user_progile_mini">
+              <li className="user_profile_mini">
                 <Link to="/main-jy">
                   <img
                     src="images/jooyoungPark/Main/mini_profile.jpg"
